Add files via upload
WebDif is for when scanning web directories based on status codes just doesn't cut it. It searches for hidden paths by matching the sequence of HTML elements or the whole page returned for each request against unwanted element sequences or whole pages. These unwanted pages might be returned with a code 200 response by a custom "not found" function, for example. They can be specified as a list of URLs in a file named "exurls" or as HTML files in a directory named "expages". Exurls and expages must be in the same directory as the main script. Run webdif, webdif -h, or webdif --help to see usage.